+++ Auth0/JWK+RSA.swift
@@ -31,7 +31,7 @@ extension JWK {
             let modulus = rsaModulus?.a0_decodeBase64URLSafe(),
             let exponent = rsaExponent?.a0_decodeBase64URLSafe() else { return nil }
         let encodedKey = encodeRSAPublicKey(modulus: [UInt8](modulus), exponent: [UInt8](exponent))
-        if #available(iOS 10, *) {
+        if #available(iOS 10.0, macOS 10.12, *) {
             return generateRSAPublicKey(from: encodedKey)
         }
         let tag = "com.auth0.tmp.RSAPublicKey"
@@ -47,7 +47,7 @@ extension JWK {
         return Data(encodedSequence)
     }
 
-    @available(iOS 10, *)
+    @available(iOS 10.0, macOS 10.12, *)
     private func generateRSAPublicKey(from derEncodedData: Data) -> SecKey? {
         let sizeInBits = derEncodedData.count * MemoryLayout<UInt8>.size
         let attributes: [CFString: Any] = [kSecClass: kSecClassKey,