+++ OAuth2Mac/ViewController.swift
@@ -7,21 +7,27 @@
 //
 
 import Cocoa
+import Auth0
 
 class ViewController: NSViewController {
 
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        // Do any additional setup after loading the view.
+    @IBAction func login(_ sender: Any) {
+        Auth0.webAuth()
+            .logging(enabled: true)
+            .start { result in
+                switch result {
+                case .success(let credentials): print(credentials)
+                case .failure(let error): print(error)
+                }
+        }
     }
-
-    override var representedObject: Any? {
-        didSet {
-        // Update the view, if already loaded.
+    
+    @IBAction func logout(_ sender: Any) {
+        Auth0.webAuth()
+            .logging(enabled: true)
+            .clearSession(federated: false) { result in
+                result ? print("Logged out") : print("Failed to log out")
         }
     }
 
-
 }
-